#139 Scripts now allows user to decide on keep the source code of dependencies (libfabric, spdlog, mutils, and etc..).

+++ scripts/prerequisites/install-libfabric.sh
@@ -16,4 +16,11 @@ libtoolize
 make -j `lscpu | grep "^CPU(" | awk '{print $2}'`
 make install
 cd ..
-rm -rf libfabric
+echo "Cleanup by executing 'rm -rf libfabric'? [Y/n]"
+read c
+if [[ $c == [nN] ]];
+then
+    echo "keep libfabric"
+else
+    rm -rf libfabric
+fi
