+++ scripts/prerequisites/install-mutils-tasks.sh
@@ -17,4 +17,12 @@ cmake -DCMAKE_BUILD_TYPE=Release -DCMAKE_INSTALL_PREFIX=${INSTALL_PREFIX} ..
 make -j `lscpu | grep "^CPU(" | awk '{print $2}'`
 make install
 cd ../..
-rm -rf mutils-tasks
+echo "Cleanup by executing 'rm -rf mutils-tasks'? [Y/n]"
+read c
+if [[ $c == [nN] ]];
+then
+    echo "keep mutils-tasks"
+else
+    rm -rf mutils-tasks
+fi
+
