+++ scripts/prerequisites/install-spdlog.sh
@@ -14,4 +14,12 @@ cmake -DCMAKE_BUILD_TYPE=Release -DCMAKE_INSTALL_PREFIX=${INSTALL_PREFIX} .
 make -j `lscpu | grep "^CPU(" | awk '{print $2}'`
 make install
 cd ..
-rm -rf spdlog
+echo "Cleanup by executing 'rm -rf spdlog'? [Y/n]"
+read c
+if [[ $c == [nN] ]];
+then
+    echo "keep spdlog"
+else
+    rm -rf spdlog
+fi
+
